requirements.txt
# Core dependencies
gradio==4.44.1
PyMuPDF==1.23.8
Pillow==10.1.0
pytesseract==0.3.10
pyttsx3==2.98
pyaudio==0.2.14
SpeechRecognition==3.10.0

# Networking & API
aiohttp==3.9.1
requests==2.31.0
aiofiles==23.2.1

# Async support
asyncio==3.4.3
async-timeout==4.0.3
aiosignal==1.3.1

# Data formats & parsing
PyYAML==6.0
python-dateutil==2.8.2

# Utilities
typing-extensions==4.9.0
python-json-logger==2.0.7
python-multipart>=0.0.9

# aiohttp dependencies
attrs==23.1.0
charset-normalizer==3.3.2
frozenlist==1.4.1
multidict==6.0.4
yarl==1.9.4

# Windows-specific dependencies
pywin32>=305.0; platform_system == "Windows"
comtypes>=1.2.0; platform_system == "Windows"

# TTS dependencies
kokoro
soundfile>=0.12.1
misaki>=0.2.0